+++ src/components/searchbar.test.tsx
@@ -1,12 +1,12 @@
+import { AppContextProvider } from '@/providers/app-context-provider';
+import { CardProviderStore } from '@/providers/';
+import { MemoryStorage } from '@/providers/storage';
 import { act, fireEvent, render, screen, waitFor } from '@testing-library/react';
-import { AppContextProvider } from 'providers/app-context-provider';
-import { CardProviderStore } from 'providers/card-provider';
 import React from 'react';
-import { MemoryStorage } from '../providers/memory-storage';
 import SearchBar from './searchbar';
 
 // Create a mock for the CardProviderStore class
-jest.mock('../providers/card-provider', () => {
+jest.mock('@/providers/', () => {
   return {
     CardProviderStore: jest.fn().mockImplementation(() => {
       return {
