+++ src/main.tsx
@@ -2,9 +2,8 @@ import App from './App';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import CardsApp from './cards-app';
-import { CardProviderStore } from './providers/card-provider';
-import { LocalStorage } from './providers/local-storage';
-import { AppContextProvider } from './providers/app-context-provider';
+import { CardProviderStore, LocalStorage } from '@/providers';
+import { AppContextProvider } from '@/providers/app-context-provider';
 
 const cardProvider = new CardProviderStore();
 const cardProvider2 = new CardProviderStore();
