+++ src/pages/former/former-page.tsx
@@ -1,8 +1,7 @@
-import React from 'react';
 import CardCreator from '@/components/card-creator';
 import CardShell from '@/components/card-shell';
-import { AppContext } from '@/providers/app-context-provider';
-import { CardData } from '@/providers/card-provider';
+import { AppContext, CardData } from '@/providers';
+import React from 'react';
 import './former-page.css';
 
 interface FormerPageState {
