+++ src/providers/storage/index.ts
@@ -1,0 +1,2 @@
+export * from './local-storage';
+export * from './memory-storage';
